iota_sdk_ffi::graphql

Module query_types

Structs

GraphQLAddress

GraphQLCoinMetadata

The coin metadata associated with the given coin type.

GraphQLDynamicFieldName

The name part of a dynamic field, including its type, bcs, and json representation.

GraphQLDynamicFieldOutput

The output of a dynamic field query, that includes the name, value, and value’s json representation.

GraphQLDynamicFieldValue

The value part of a dynamic field.

GraphQLEpoch

GraphQLEvent

An event as returned by the GraphQL events query.

GraphQLEventFilter

GraphQLMoveEnum

A Move enum definition.

GraphQLMoveEnumConnection

A page of Move enum definitions.

GraphQLMoveEnumVariant

A variant of a Move enum.

GraphQLMoveField

A field of a Move struct or enum variant.

GraphQLMoveFunction

GraphQLMoveFunctionConnection

GraphQLMoveFunctionTypeParameter

A type parameter of a Move function, with the abilities it is constrained to.

GraphQLMoveModule

GraphQLMoveModuleConnection

GraphQLMoveModuleQuery

GraphQLMoveObject

The BCS contents of an on-chain Move object.

GraphQLMoveStructConnection

A page of Move struct definitions.

GraphQLMoveStructQuery

A Move struct definition.

GraphQLMoveStructTypeParameter

A type parameter of a Move struct, with the abilities it is constrained to.

GraphQLMoveViewResult

The result of executing a Move View Function.

GraphQLObjectFilter

GraphQLObjectRef

GraphQLOpenMoveType

A Move type that may still have unbound type parameters.

GraphQLPageInfo

Information about pagination in a connection.

GraphQLPaginationFilter

Pagination options for querying the GraphQL server. It defaults to forward pagination with the GraphQL server’s max page size.

GraphQLProtocolConfigAttr

A key-value protocol configuration attribute.

GraphQLProtocolConfigFeatureFlag

Feature flags are a form of boolean configuration that are usually used to gate features while they are in development. Once a lag has been enabled, it is rare for it to be disabled.

GraphQLProtocolConfigs

Information about the configuration of the protocol. Constants that control how the chain operates. These can only change during protocol upgrades which happen on epoch boundaries.

GraphQLServiceConfig

GraphQLTransactionDataEffects

GraphQLTransactionMetadata

GraphQLTransactionsFilter

Filter for transaction queries.

GraphQLValidator

Represents a validator in the system.

GraphQLValidatorConnection

GraphQLValidatorCredentials

The credentials related fields associated with a validator.

GraphQLValidatorSet

Enums

GraphQLAddressTransactionRelationship

The relationship between an address and a transaction.

GraphQLDirection

Pagination direction.

GraphQLFeature

A feature an RPC service can have enabled.

GraphQLMoveAbility

An ability a Move type can have.

GraphQLMoveVisibility

The visibility of a Move function.

GraphQLTransactionBlockKindInput

The kind of a transaction, used to filter transaction queries.