iota_sdk_ffi::graphql

Struct GraphQLClient

pub struct GraphQLClient(/* private fields */);

The GraphQL client for interacting with the IOTA blockchain.

Implementations

impl GraphQLClient

pub fn inner(&self) -> &RwLock<Client>

pub fn into_inner(self) -> RwLock<Client>

impl GraphQLClient

pub fn new(server: String) -> Result<Self>

Create a new GraphQL client with the provided server address.

pub fn new_mainnet() -> Self

Create a new GraphQL client connected to the mainnet GraphQL server: {MAINNET_HOST}.

pub fn new_testnet() -> Self

Create a new GraphQL client connected to the testnet GraphQL server: {TESTNET_HOST}.

pub fn new_devnet() -> Self

Create a new GraphQL client connected to the devnet GraphQL server: {DEVNET_HOST}.

pub fn new_localnet() -> Self

Create a new GraphQL client connected to the localhost GraphQL server: {DEFAULT_LOCAL_HOST}.

pub async fn chain_id(&self) -> Result<String>

Get the chain identifier.

pub async fn max_page_size(&self) -> Result<i32>

Lazily fetch the max page size

pub async fn reference_gas_price( &self, epoch: Option<u64>, ) -> Result<Option<u64>>

Get the reference gas price for the provided epoch or the last known one if no epoch is provided.

This will return Ok(None) if the epoch requested is not available in the GraphQL service (e.g., due to pruning).

pub async fn protocol_config( &self, version: Option<u64>, ) -> Result<ProtocolConfigs>

Get the protocol configuration.

pub async fn active_validators( &self, epoch: Option<u64>, pagination_filter: Option<PaginationFilter>, ) -> Result<ValidatorPage>

Get the list of active validators for the provided epoch, including related metadata. If no epoch is provided, it will return the active validators for the current epoch.

pub async fn total_transaction_blocks_by_digest( &self, digest: &Digest, ) -> Result<Option<u64>>

The total number of transaction blocks in the network by the end of the provided checkpoint digest.

pub async fn total_transaction_blocks_by_seq_num( &self, seq_num: u64, ) -> Result<Option<u64>>

The total number of transaction blocks in the network by the end of the provided checkpoint sequence number.

pub async fn total_transaction_blocks(&self) -> Result<Option<u64>>

The total number of transaction blocks in the network by the end of the last known checkpoint.

pub async fn coins( &self, owner: &Address, pagination_filter: Option<PaginationFilter>, coin_type: Option<Arc<StructTag>>, ) -> Result<CoinPage>

Get the list of coins for the specified address.

If coin_type is not provided, all coins will be returned. For IOTA coins, pass in the coin type: 0x2::iota::IOTA.

pub async fn gas_coins( &self, owner: &Address, pagination_filter: Option<PaginationFilter>, ) -> Result<CoinPage>

Get the list of gas coins for the specified address.

pub async fn coin_metadata( &self, coin_type: &str, ) -> Result<Option<CoinMetadata>>

Get the coin metadata for the coin type.

pub async fn total_supply(&self, coin_type: &str) -> Result<Option<u64>>

Get total supply for the coin type.

pub async fn checkpoint( &self, digest: Option<Arc<Digest>>, seq_num: Option<u64>, ) -> Result<Option<Arc<CheckpointSummary>>>

Get the CheckpointSummary for a given checkpoint digest or checkpoint id. If none is provided, it will use the last known checkpoint id.

pub async fn checkpoints( &self, pagination_filter: Option<PaginationFilter>, ) -> Result<CheckpointSummaryPage>

Get a page of CheckpointSummary for the provided parameters.

pub async fn latest_checkpoint_sequence_number( &self, ) -> Result<Option<CheckpointSequenceNumber>>

Return the sequence number of the latest checkpoint that has been executed.

pub async fn epoch(&self, epoch: Option<u64>) -> Result<Option<Epoch>>

Return the epoch information for the provided epoch. If no epoch is provided, it will return the last known epoch.

pub async fn epoch_total_checkpoints( &self, epoch: Option<u64>, ) -> Result<Option<u64>>

Return the number of checkpoints in this epoch. This will return Ok(None) if the epoch requested is not available in the GraphQL service (e.g., due to pruning).

pub async fn epoch_total_transaction_blocks( &self, epoch: Option<u64>, ) -> Result<Option<u64>>

Return the number of transaction blocks in this epoch. This will return Ok(None) if the epoch requested is not available in the GraphQL service (e.g., due to pruning).

pub async fn events( &self, filter: Option<EventFilter>, pagination_filter: Option<PaginationFilter>, ) -> Result<EventPage>

Return a page of tuple (event, transaction digest) based on the (optional) event filter.

pub async fn object( &self, object_id: &ObjectId, version: Option<u64>, ) -> Result<Option<Arc<Object>>>

Return an object based on the provided Address.

If the object does not exist (e.g., due to pruning), this will return Ok(None). Similarly, if this is not an object but an address, it will return Ok(None).

pub async fn objects( &self, filter: Option<ObjectFilter>, pagination_filter: Option<PaginationFilter>, ) -> Result<ObjectPage>

Return a page of objects based on the provided parameters.

Use this function together with the ObjectFilter::owner to get the objects owned by an address.

pub async fn object_bcs(&self, object_id: &ObjectId) -> Result<Option<Vec<u8>>>

Return the object’s bcs content Vec<u8> based on the provided Address.

pub async fn move_object_contents_bcs( &self, object_id: &ObjectId, version: Option<u64>, ) -> Result<Option<Vec<u8>>>

Return the BCS of an object that is a Move object.

If the object does not exist (e.g., due to pruning), this will return Ok(None). Similarly, if this is not an object but an address, it will return Ok(None).

pub async fn package( &self, address: &Address, version: Option<u64>, ) -> Result<Option<Arc<MovePackage>>>

The package corresponding to the given address (at the optionally given version). When no version is given, the package is loaded directly from the address given. Otherwise, the address is translated before loading to point to the package whose original ID matches the package at address, but whose version is version. For non-system packages, this might result in a different address than address because different versions of a package, introduced by upgrades, exist at distinct addresses.

Note that this interpretation of version is different from a historical object read (the interpretation of version for the object query).

pub async fn package_versions( &self, address: &Address, after_version: Option<u64>, before_version: Option<u64>, pagination_filter: Option<PaginationFilter>, ) -> Result<MovePackagePage>

Fetch all versions of package at address (packages that share this package’s original ID), optionally bounding the versions exclusively from below with afterVersion, or from above with beforeVersion.

pub async fn package_latest( &self, address: &Address, ) -> Result<Option<Arc<MovePackage>>>

Fetch the latest version of the package at address. This corresponds to the package with the highest version that shares its original ID with the package at address.

pub async fn packages( &self, after_checkpoint: Option<u64>, before_checkpoint: Option<u64>, pagination_filter: Option<PaginationFilter>, ) -> Result<MovePackagePage>

The Move packages that exist in the network, optionally filtered to be strictly before beforeCheckpoint and/or strictly after afterCheckpoint.

This query returns all versions of a given user package that appear between the specified checkpoints, but only records the latest versions of system packages.

pub async fn transaction( &self, digest: &Digest, ) -> Result<Option<SignedTransaction>>

Get a transaction by its digest.

pub async fn transaction_effects( &self, digest: &Digest, ) -> Result<Option<Arc<TransactionEffects>>>

Get a transaction’s effects by its digest.

pub async fn transaction_data_effects( &self, digest: &Digest, ) -> Result<Option<TransactionDataEffects>>

Get a transaction’s data and effects by its digest.

pub async fn transactions( &self, filter: Option<TransactionsFilter>, pagination_filter: Option<PaginationFilter>, ) -> Result<SignedTransactionPage>

Get a page of transactions based on the provided filters.

pub async fn transactions_effects( &self, filter: Option<TransactionsFilter>, pagination_filter: Option<PaginationFilter>, ) -> Result<TransactionEffectsPage>

Get a page of transactions’ effects based on the provided filters.

pub async fn transactions_data_effects( &self, filter: Option<TransactionsFilter>, pagination_filter: Option<PaginationFilter>, ) -> Result<TransactionDataEffectsPage>

Get a page of transactions’ data and effects based on the provided filters.

pub async fn execute_tx( &self, signatures: Vec<Arc<UserSignature>>, tx: &Transaction, wait_for: Option<WaitForTx>, ) -> Result<TransactionEffects>

Execute a transaction.

pub async fn is_tx_indexed_on_node(&self, digest: &Digest) -> Result<bool>

Returns whether the transaction for the given digest has been indexed on the node. This means that it can be queries by its digest and its effects will be usable for subsequent transactions. To check for full finalization, use is_tx_finalized.

pub async fn is_tx_finalized(&self, digest: &Digest) -> Result<bool>

Returns whether the transaction for the given digest has been included in a checkpoint (finalized).

pub async fn wait_for_tx( &self, digest: &Digest, wait_for: WaitForTx, timeout: Option<Duration>, ) -> Result<()>

Wait for the indexing or finalization of a transaction by its digest. An optional timeout can be provided, which, if exceeded, will return an error (default 60s).

pub async fn normalized_move_function( &self, package: &Address, module: &str, function: &str, version: Option<u64>, ) -> Result<Option<Arc<MoveFunction>>>

Return the normalized Move function data for the provided package, module, and function.

pub async fn move_object_contents( &self, object_id: &ObjectId, version: Option<u64>, ) -> Result<Option<Value>>

Return the contents’ JSON of an object that is a Move object.

If the object does not exist (e.g., due to pruning), this will return Ok(None). Similarly, if this is not an object but an address, it will return Ok(None).

pub async fn normalized_move_module( &self, package: &Address, module: &str, version: Option<u64>, pagination_filter_enums: Option<PaginationFilter>, pagination_filter_friends: Option<PaginationFilter>, pagination_filter_functions: Option<PaginationFilter>, pagination_filter_structs: Option<PaginationFilter>, ) -> Result<Option<MoveModule>>

Return the normalized Move module data for the provided module.

pub async fn dynamic_field( &self, address: &Address, type_tag: &TypeTag, name: Value, ) -> Result<Option<DynamicFieldOutput>>

Access a dynamic field on an object using its name. Names are arbitrary Move values whose type have copy, drop, and store, and are specified using their type, and their BCS contents, Base64 encoded.

The name argument is a json serialized type.

This returns DynamicFieldOutput which contains the name, the value as json, and object.

pub async fn dynamic_object_field( &self, address: &Address, type_tag: &TypeTag, name: Value, ) -> Result<Option<DynamicFieldOutput>>

Access a dynamic object field on an object using its name. Names are arbitrary Move values whose type have copy, drop, and store, and are specified using their type, and their BCS contents, Base64 encoded.

The name argument is a json serialized type.

This returns DynamicFieldOutput which contains the name, the value as json, and object.

pub async fn dynamic_fields( &self, address: &Address, pagination_filter: Option<PaginationFilter>, ) -> Result<DynamicFieldOutputPage>

Get a page of dynamic fields for the provided address. Note that this will also fetch dynamic fields on wrapped objects.

This returns a page of DynamicFieldOutputs.

pub async fn set_rpc_server(&self, server: String) -> Result<()>

Set the server address for the GraphQL GraphQL client. It should be a valid URL with a host and optionally a port number.

pub async fn service_config(&self) -> Result<ServiceConfig>

Get the GraphQL service configuration, including complexity limits, read and mutation limits, supported versions, and others.

pub async fn dry_run_tx( &self, tx: &Transaction, skip_checks: bool, ) -> Result<DryRunResult>

Dry run a Transaction and return the transaction effects and dry run error (if any).

skipChecks optional flag disables the usual verification checks that prevent access to objects that are owned by addresses other than the sender, and calling non-public, non-entry functions, and some other checks. Defaults to false.

pub async fn dry_run_tx_kind( &self, tx_kind: &TransactionKind, tx_meta: TransactionMetadata, skip_checks: bool, ) -> Result<DryRunResult>

Dry run a TransactionKind and return the transaction effects and dry run error (if any).

skipChecks optional flag disables the usual verification checks that prevent access to objects that are owned by addresses other than the sender, and calling non-public, non-entry functions, and some other checks. Defaults to false.

tx_meta is the transaction metadata.

pub async fn run_query(&self, query: Query) -> Result<Value>

Run a query.

pub async fn balance( &self, address: &Address, coin_type: Option<String>, ) -> Result<Option<u64>>

Get the balance of all the coins owned by address for the provided coin type. Coin type will default to 0x2::coin::Coin<0x2::iota::IOTA> if not provided.

pub async fn iota_names_lookup( &self, name: &str, ) -> Result<Option<Arc<Address>>>

Return the resolved address for the given name.

pub async fn iota_names_registrations( &self, address: &Address, pagination_filter: PaginationFilter, ) -> Result<NameRegistrationPage>

Find all registration NFTs for the given address.

pub async fn iota_names_default_name( &self, address: &Address, format: Option<NameFormat>, ) -> Result<Option<Arc<Name>>>

Get the default name pointing to this address, if one exists.