iota_sdk_ffi::types::crypto

Module multisig

Structs

MultisigAggregatedSignature

Aggregated signature from members of a multisig committee.

MultisigCommittee

A multisig committee

MultisigMember

A member in a multisig committee

MultisigMemberPublicKey

Enum of valid public keys for multisig committee members

MultisigMemberSignature

A signature from a member of a multisig committee.

Functions

multisig_aggregated_signature_from_bcs

Create this type from BCS encoded bytes.

multisig_aggregated_signature_to_bcs

Convert this type to BCS encoded bytes.

multisig_committee_from_bcs

Create this type from BCS encoded bytes.

multisig_committee_to_bcs

Convert this type to BCS encoded bytes.

multisig_member_from_bcs

Create this type from BCS encoded bytes.

multisig_member_public_key_from_bcs

Create this type from BCS encoded bytes.

multisig_member_public_key_to_bcs

Convert this type to BCS encoded bytes.

multisig_member_signature_from_bcs

Create this type from BCS encoded bytes.

multisig_member_signature_to_bcs

Convert this type to BCS encoded bytes.

multisig_member_to_bcs

Convert this type to BCS encoded bytes.