iota_sdk_ffi::types

Module graphql

Structs

CoinMetadata

The coin metadata associated with the given coin type.

DynamicFieldName

The name part of a dynamic field, including its type, bcs, and json representation.

DynamicFieldOutput

The output of a dynamic field query, that includes the name, value, and value’s json representation.

DynamicFieldValue

The value part of a dynamic field.

Epoch

EventFilter

GQLAddress

MoveFunction

MoveFunctionConnection

MoveModule

MoveModuleConnection

MoveModuleQuery

MovePackageQuery

ObjectFilter

ObjectRef

TransactionDataEffects

TransactionMetadata

TransactionsFilter

Validator

Represents a validator in the system.

ValidatorConnection

ValidatorSet