iota_sdk_ffi::types

Module object

Structs

GenesisObject

An object part of the initial chain state

MovePackage

A move package

MoveStruct

A move struct

Object

An object on the IOTA blockchain

ObjectData

Object data, either a package or struct

ObjectId

An ObjectId is a 32-byte identifier used to uniquely identify an object on the IOTA blockchain.

ObjectReference

Reference to an object

ObjectType

Type of an IOTA object

Owner

Enum of different types of ownership for an object.

TypeOrigin

Identifies a struct and the module it was defined in

UpgradeInfo

Upgraded package info for the linkage table

Functions

genesis_object_from_bcs

Create this type from BCS encoded bytes.

genesis_object_to_bcs

Convert this type to BCS encoded bytes.

move_package_from_bcs

Create this type from BCS encoded bytes.

move_package_to_bcs

Convert this type to BCS encoded bytes.

move_struct_from_bcs

Create this type from BCS encoded bytes.

move_struct_to_bcs

Convert this type to BCS encoded bytes.

object_data_from_bcs

Create this type from BCS encoded bytes.

object_data_to_bcs

Convert this type to BCS encoded bytes.

object_from_bcs

Create this type from BCS encoded bytes.

object_id_from_bcs

Create this type from BCS encoded bytes.

object_id_to_bcs

Convert this type to BCS encoded bytes.

object_reference_from_bcs

Create this type from BCS encoded bytes.

object_reference_to_bcs

Convert this type to BCS encoded bytes.

object_to_bcs

Convert this type to BCS encoded bytes.

owner_from_bcs

Create this type from BCS encoded bytes.

owner_to_bcs

Convert this type to BCS encoded bytes.

type_origin_from_bcs

Create this type from BCS encoded bytes.

type_origin_to_bcs

Convert this type to BCS encoded bytes.

upgrade_info_from_bcs

Create this type from BCS encoded bytes.

upgrade_info_to_bcs

Convert this type to BCS encoded bytes.

Type Aliases

Version