iota_sdk_ffi::graphql

Module query_types

Structs

CoinMetadata

The coin metadata associated with the given coin type.

DynamicFieldName

The name part of a dynamic field, including its type, bcs, and json representation.

DynamicFieldOutput

The output of a dynamic field query, that includes the name, value, and value’s json representation.

DynamicFieldValue

The value part of a dynamic field.

Epoch

EventFilter

GraphQLAddress

GraphQLEvent

An event as returned by the GraphQL events query.

MoveEnum

A Move enum definition.

MoveEnumConnection

A page of Move enum definitions.

MoveEnumVariant

A variant of a Move enum.

MoveField

A field of a Move struct or enum variant.

MoveFunction

MoveFunctionConnection

MoveFunctionTypeParameter

A type parameter of a Move function, with the abilities it is constrained to.

MoveModule

MoveModuleConnection

MoveModuleQuery

MoveObject

The BCS contents of an on-chain Move object.

MovePackageQuery

MoveStructConnection

A page of Move struct definitions.

MoveStructQuery

A Move struct definition.

MoveStructTypeParameter

A type parameter of a Move struct, with the abilities it is constrained to.

MoveViewResult

The result of executing a Move View Function.

ObjectFilter

ObjectRef

OpenMoveType

A Move type that may still have unbound type parameters.

PageInfo

Information about pagination in a connection.

PaginationFilter

Pagination options for querying the GraphQL server. It defaults to forward pagination with the GraphQL server’s max page size.

ProtocolConfigAttr

A key-value protocol configuration attribute.

ProtocolConfigFeatureFlag

Feature flags are a form of boolean configuration that are usually used to gate features while they are in development. Once a lag has been enabled, it is rare for it to be disabled.

ProtocolConfigs

Information about the configuration of the protocol. Constants that control how the chain operates. These can only change during protocol upgrades which happen on epoch boundaries.

ServiceConfig

TransactionDataEffects

TransactionMetadata

TransactionsFilter

Validator

Represents a validator in the system.

ValidatorConnection

ValidatorCredentials

The credentials related fields associated with a validator.

ValidatorSet

Enums

Direction

Pagination direction.

Feature

A feature an RPC service can have enabled.

MoveAbility

An ability a Move type can have.

MoveVisibility

The visibility of a Move function.

TransactionBlockKindInput

The kind of a transaction, used to filter transaction queries.