iota_sdk_ffi::types

Module transaction

Modules

v1

Structs

Argument

An argument to a programmable transaction command

CanceledTransaction

A transaction that was canceled

ChangeEpoch

System transaction used to change the epoch

ChangeEpochV2

System transaction used to change the epoch

ChangeEpochV3

ChangeEpochV4

Command

A single command in a programmable transaction.

ConsensusCommitPrologueV1

V1 of the consensus commit prologue system transaction

ConsensusDeterminedVersionAssignments

DenyRuleSet

A complete set of transaction deny rules.

EndOfEpochTransactionKind

Operation run at the end of an epoch

GasPayment

Payment information for executing a transaction

GenesisTransaction

The genesis transaction

Input

An input to a user transaction

MakeMoveVector

Command to build a move vector out of a set of individual elements

MergeCoins

Command to merge multiple coins of the same type into a single coin

MoveCall

Command to call a move function

ProgrammableTransaction

A user transaction

Publish

Command to publish a new move package

RandomnessStateUpdate

Randomness update

SharedObjectReference

A shared object input to a programmable transaction

SignedTransaction

SplitCoins

Command to split a single coin object into multiple coins

SystemPackage

System package

Transaction

Transaction

TransactionDenyRulesUpdate

Update of the on-chain transaction deny rules.

TransactionEffects

The output or effects of executing a transaction

TransactionKind

Transaction type

TransactionV1

A transaction

TransferObjects

Command to transfer ownership of a set of objects to an address

Upgrade

Command to upgrade an already published package

VersionAssignment

Object version assignment from consensus

Enums

TransactionExpiration

A TTL for a transaction

Functions

deny_rule_set_from_bcs

Create this type from BCS encoded bytes.

deny_rule_set_from_json

Create this type from a JSON encoded string.

gas_payment_from_bcs

Create this type from BCS encoded bytes.

gas_payment_from_json

Create this type from a JSON encoded string.

randomness_state_update_from_bcs

Create this type from BCS encoded bytes.

randomness_state_update_from_json

Create this type from a JSON encoded string.

signed_transaction_from_bcs

Create this type from BCS encoded bytes.

signed_transaction_from_json

Create this type from a JSON encoded string.

transaction_deny_rules_update_from_bcs

Create this type from BCS encoded bytes.

transaction_deny_rules_update_from_json

Create this type from a JSON encoded string.

transaction_expiration_from_bcs

Create this type from BCS encoded bytes.

transaction_expiration_from_json

Create this type from a JSON encoded string.