iota_sdk_ffi::types

Module transaction

Modules

v1

Structs

Argument

An argument to a programmable transaction command

CancelledTransaction

A transaction that was cancelled

ChangeEpoch

System transaction used to change the epoch

ChangeEpochV2

System transaction used to change the epoch

ChangeEpochV3

Command

A single command in a programmable transaction.

ConsensusCommitPrologueV1

V1 of the consensus commit prologue system transaction

ConsensusDeterminedVersionAssignments

DryRunEffect

Effects of a single command in the dry run, including mutated references and return values.

DryRunMutation

A mutation to an argument that was mutably borrowed by a command.

DryRunResult

The result of a simulation (dry run), which includes the effects of the transaction, any errors that may have occurred, and intermediate results for each command.

DryRunReturn

A return value from a command in the dry run.

EndOfEpochTransactionKind

Operation run at the end of an epoch

ExecutionTimeObservation

ExecutionTimeObservationKey

Key for an execution time observation

ExecutionTimeObservations

Set of Execution Time Observations from the committee.

GasPayment

Payment information for executing a transaction

GenesisTransaction

The genesis transaction

Input

An input to a user transaction

MakeMoveVector

Command to build a move vector out of a set of individual elements

MergeCoins

Command to merge multiple coins of the same type into a single coin

MoveCall

Command to call a move function

ProgrammableTransaction

A user transaction

Publish

Command to publish a new move package

SignedTransaction

SplitCoins

Command to split a single coin object into multiple coins

SystemPackage

System package

Transaction

Transaction

TransactionEffects

The output or effects of executing a transaction

TransactionKind

Transaction type

TransactionV1

A transaction

TransferObjects

Command to transfer ownership of a set of objects to an address

Upgrade

Command to upgrade an already published package

ValidatorExecutionTimeObservation

An execution time observation from a particular validator

VersionAssignment

Object version assignment from consensus

Enums

TransactionArgument

A transaction argument used in programmable transactions.

Functions

active_jwk_from_bcs

Create this type from BCS encoded bytes.

active_jwk_to_bcs

Convert this type to BCS encoded bytes.

argument_from_bcs

Create this type from BCS encoded bytes.

argument_to_bcs

Convert this type to BCS encoded bytes.

authenticator_state_expire_from_bcs

Create this type from BCS encoded bytes.

authenticator_state_expire_to_bcs

Convert this type to BCS encoded bytes.

authenticator_state_update_v1_from_bcs

Create this type from BCS encoded bytes.

authenticator_state_update_v1_to_bcs

Convert this type to BCS encoded bytes.

cancelled_transaction_from_bcs

Create this type from BCS encoded bytes.

cancelled_transaction_to_bcs

Convert this type to BCS encoded bytes.

change_epoch_from_bcs

Create this type from BCS encoded bytes.

change_epoch_to_bcs

Convert this type to BCS encoded bytes.

change_epoch_v2_from_bcs

Create this type from BCS encoded bytes.

change_epoch_v2_to_bcs

Convert this type to BCS encoded bytes.

command_from_bcs

Create this type from BCS encoded bytes.

command_to_bcs

Convert this type to BCS encoded bytes.

consensus_commit_prologue_v1_from_bcs

Create this type from BCS encoded bytes.

consensus_commit_prologue_v1_to_bcs

Convert this type to BCS encoded bytes.

consensus_determined_version_assignments_from_bcs

Create this type from BCS encoded bytes.

consensus_determined_version_assignments_to_bcs

Convert this type to BCS encoded bytes.

execution_time_observation_from_bcs

Create this type from BCS encoded bytes.

execution_time_observation_key_from_bcs

Create this type from BCS encoded bytes.

execution_time_observation_key_to_bcs

Convert this type to BCS encoded bytes.

execution_time_observation_to_bcs

Convert this type to BCS encoded bytes.

execution_time_observations_from_bcs

Create this type from BCS encoded bytes.

execution_time_observations_to_bcs

Convert this type to BCS encoded bytes.

gas_payment_from_bcs

Create this type from BCS encoded bytes.

gas_payment_to_bcs

Convert this type to BCS encoded bytes.

genesis_transaction_from_bcs

Create this type from BCS encoded bytes.

genesis_transaction_to_bcs

Convert this type to BCS encoded bytes.

input_from_bcs

Create this type from BCS encoded bytes.

input_to_bcs

Convert this type to BCS encoded bytes.

make_move_vector_from_bcs

Create this type from BCS encoded bytes.

make_move_vector_to_bcs

Convert this type to BCS encoded bytes.

merge_coins_from_bcs

Create this type from BCS encoded bytes.

merge_coins_to_bcs

Convert this type to BCS encoded bytes.

move_call_from_bcs

Create this type from BCS encoded bytes.

move_call_to_bcs

Convert this type to BCS encoded bytes.

programmable_transaction_from_bcs

Create this type from BCS encoded bytes.

programmable_transaction_to_bcs

Convert this type to BCS encoded bytes.

publish_from_bcs

Create this type from BCS encoded bytes.

publish_to_bcs

Convert this type to BCS encoded bytes.

randomness_state_update_from_bcs

Create this type from BCS encoded bytes.

randomness_state_update_to_bcs

Convert this type to BCS encoded bytes.

signed_transaction_from_bcs

Create this type from BCS encoded bytes.

signed_transaction_to_bcs

Convert this type to BCS encoded bytes.

split_coins_from_bcs

Create this type from BCS encoded bytes.

split_coins_to_bcs

Convert this type to BCS encoded bytes.

system_package_from_bcs

Create this type from BCS encoded bytes.

system_package_to_bcs

Convert this type to BCS encoded bytes.

transaction_effects_from_bcs

Create this type from BCS encoded bytes.

transaction_effects_to_bcs

Convert this type to BCS encoded bytes.

transaction_expiration_from_bcs

Create this type from BCS encoded bytes.

transaction_expiration_to_bcs

Convert this type to BCS encoded bytes.

transaction_from_bcs

Create this type from BCS encoded bytes.

transaction_kind_from_bcs

Create this type from BCS encoded bytes.

transaction_kind_to_bcs

Convert this type to BCS encoded bytes.

transaction_to_bcs

Convert this type to BCS encoded bytes.

transaction_v1_from_bcs

Create this type from BCS encoded bytes.

transaction_v1_to_bcs

Convert this type to BCS encoded bytes.

transfer_objects_from_bcs

Create this type from BCS encoded bytes.

transfer_objects_to_bcs

Convert this type to BCS encoded bytes.

upgrade_from_bcs

Create this type from BCS encoded bytes.

upgrade_to_bcs

Convert this type to BCS encoded bytes.

validator_execution_time_observation_from_bcs

Create this type from BCS encoded bytes.

validator_execution_time_observation_to_bcs

Convert this type to BCS encoded bytes.

version_assignment_from_bcs

Create this type from BCS encoded bytes.

version_assignment_to_bcs

Convert this type to BCS encoded bytes.